+++ src/main/java/org/opentripplanner/api/common/RoutingResource.java
@@ -39,7 +39,8 @@ import org.slf4j.LoggerFactory;
  *
  * @author abyrd
  */
-public abstract class RoutingResource { 
+@SuppressWarnings("FieldMayBeFinal")
+public abstract class RoutingResource {
 
     private static final Logger LOG = LoggerFactory.getLogger(RoutingResource.class);
 
@@ -709,10 +710,9 @@ public abstract class RoutingResource {
     /**
      * Range/sanity check the query parameter fields and build a Request object from them.
      *
-     * @throws ParameterException when there is a problem interpreting a query parameter
      * @param queryParameters incoming request parameters
      */
-    protected RoutingRequest buildRequest(MultivaluedMap<String, String> queryParameters) throws ParameterException {
+    protected RoutingRequest buildRequest(MultivaluedMap<String, String> queryParameters) {
         Router router = otpServer.getRouter();
         RoutingRequest request = router.copyDefaultRoutingRequest();