+++ src/main/java/org/opentripplanner/api/mapping/StopMapper.java
@@ -3,7 +3,6 @@ package org.opentripplanner.api.mapping;
 import java.util.Collection;
 import java.util.List;
 import java.util.stream.Collectors;
-
 import org.opentripplanner.api.model.ApiStop;
 import org.opentripplanner.api.model.ApiStopShort;
 import org.opentripplanner.model.Stop;
@@ -32,7 +31,7 @@ public class StopMapper {
         if (extended) {
             api.desc = domain.getDescription();
             api.zoneId = domain.getFirstZoneAsString();
-            api.url = domain.getUrl().toString();
+            api.url = I18NStringMapper.mapToApi(domain.getUrl(), null);
             api.locationType = 0;
             api.stationId = FeedScopedIdMapper.mapIdToApi(domain.getParentStation());
             api.parentStation = mapToParentStationOldId(domain);
@@ -54,7 +53,7 @@ public class StopMapper {
         api.name = domain.getName().toString();
         api.lat = domain.getLat();
         api.lon = domain.getLon();
-        api.url = domain.getUrl().toString();
+        api.url = I18NStringMapper.mapToApi(domain.getUrl(), null);
         api.stationId = FeedScopedIdMapper.mapIdToApi(domain.getParentStation());
         // parentStation may be missing on the stop returning null.
         // TODO harmonize these names, maybe use "station" everywhere