+++ src/main/java/org/opentripplanner/util/MapUtils.java
@@ -22,56 +22,6 @@ public class MapUtils {
         set.add(value);
     }
 
-    public static <T, U> void addToMapList(Map<T, List<U>> mapList, T key, U value) {
-        List<U> list = mapList.get(key);
-        if (list == null) {
-            list = new ArrayList<>();
-            mapList.put(key, list);
-        }
-        list.add(value);
-    }
-
-    public static <T> boolean addToMaxMap(Map<T, Double> map, T key, double value) {
-        Double oldValue = map.get(key);
-        if (oldValue == null || value > oldValue) {
-            map.put(key, value);
-            return true;
-        }
-        return false;
-    }
-
-    public static <T, U> void addToMapListUnique(Map<T, List<U>> mapList,
-            T key, List<U> values) {
-        List<U> list = mapList.get(key);
-        if (list == null) {
-            list = new ArrayList<>(values.size());
-            mapList.put(key, list);
-        }
-        for (U value : values) {
-            if (!list.contains(value)) {
-                list.add(value);
-            }
-        }
-    }
-
-    public static <T, U, V extends Collection<U>> void mergeInUnique(Map<T, V> mapList,
-            Map<T, V> from) {
-        for (Map.Entry<T, V> entry : from.entrySet()) {
-            T key = entry.getKey();
-            V value = entry.getValue();
-            V originalValue = mapList.get(key);
-            if (originalValue != null) {
-                HashSet<U> originalSet = new HashSet<>(originalValue);
-                for (U item : value) {
-                    if (!originalSet.contains(item))
-                        originalValue.add(item);
-                }
-            } else {
-                mapList.put(key, value);
-            }
-        }
-    }
-
     /**
      * Map a collection of objects of type <em>S</em> to a list of type <em>T</em> using the
      * provided mapping function.