+++ src/test/java/org/opentripplanner/routing/edgetype/TriangleTest.java
@@ -4,7 +4,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
-import org.opentripplanner.common.geometry.PackedCoordinateSequence;
+import org.locationtech.jts.geom.impl.PackedCoordinateSequence;
 import org.opentripplanner.routing.core.BicycleOptimizeType;
 import org.opentripplanner.routing.core.State;
 import org.opentripplanner.routing.core.TraverseMode;